chore(main): migrate deprecated code to new main
