+++ internal/deprecated/control-plane_test.go
@@ -1,19 +1,16 @@
-package main
+package deprecated
 
 import (
 	"fmt"
 	"testing"
 	"time"
 
-	"github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/assert"
 
 	"github.com/newrelic/nri-kubernetes/v2/src/apiserver"
 	"github.com/newrelic/nri-kubernetes/v2/src/controlplane"
 	"github.com/newrelic/nri-kubernetes/v2/src/definition"
 )
-
-var logger = logrus.StandardLogger()
 
 func TestControlPlaneJobs(t *testing.T) {
 	nodeName := "ip-10.0.2.15"
